+++ backend/canisters/user_index/api/src/updates/submit_proof_of_unique_personhood.rs
@@ -1,8 +1,9 @@
-use candid::CandidType;
+use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct Args {
+    pub user_ii_principal: Principal,
     pub credential_jwt: String,
 }