+++ backend/canisters/user_index/impl/src/updates/submit_proof_of_unique_personhood.rs
@@ -21,7 +21,7 @@ fn submit_proof_of_unique_personhood_impl(args: Args, state: &mut RuntimeState) 
 
     let now = state.env.now();
     match verify_proof_of_unique_personhood(
-        caller,
+        args.user_ii_principal,
         state.data.internet_identity_canister_id,
         &args.credential_jwt,
         &state.data.ic_root_key,